Create a states system
This is system is very simple rn and will expand as I get more anims. Currently it reads 7 states and applies animations based on the player's stats. Its current very basic and achieves this by reading the currentlyPlaying variable. Depending on its value when the state is changed it will play an animation. Took about 2 hours.